Move logic over to streamer
